Reviewed pull request for reverse commands
